Refactor: split core logic into shared library and CLI tool
