+++ src/core/odin_types.h
@@ -15,45 +15,6 @@
 enum class ExitCode : int { Success = 0, Usage = 2, Usb = 3, Firmware = 4, Pit = 5, Protocol = 6 };
 
 // ============================================================================
-// CONFIGURATION STRUCTURES
-// ============================================================================
-
-struct OdinConfig {
-    std::string bootloader;
-    std::string ap;
-    std::string cp;
-    std::string csc;
-    std::string ums;
-    std::string device_path;
-    // Flags controlling optional behaviour
-    bool reboot = false;
-    bool redownload = false;
-    // Perform a dry run (check-only) instead of actually flashing. When
-    // enabled, odin4 will verify firmware integrity and compatibility,
-    // parse PIT tables and tar archives, but will not send any data to
-    // the device. This flag corresponds to the --check-only command-line
-    // option.
-    bool dry_run = false;
-
-    // When enabled, files that do not map to PIT partitions are skipped instead
-    // of causing an abort. This is disabled by default for safety.
-    bool allow_unknown = false;
-
-    // Logging
-    bool quiet = false;
-    bool verbose = false;
-    bool debug = false;
-
-    // Optional USB selection overrides.
-    bool has_vid = false;
-    uint16_t vid = 0;
-    bool has_pid = false;
-    uint16_t pid = 0;
-    bool has_usb_interface = false;
-    int usb_interface = 0;
-};
-
-// ============================================================================
 // PARTITION INFORMATION TABLE (PIT)
 // ============================================================================
